requirements.txt
python-telegram-bot==20.8
firebase-admin==6.5.0
python-dotenv==1.0.1